// src/generated/models/Cre2b_modulosModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cre2b_modulosstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cre2b_modulosstatecode = keyof typeof Cre2b_modulosstatecode;
export const Cre2b_modulosstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cre2b_modulosstatuscode = keyof typeof Cre2b_modulosstatuscode;

export interface Cre2b_modulosBase {
  "cre2b_Aplicacion@odata.bind"?: string;
  cre2b_moduloid: string;
  cre2b_nombre?: string;
  cre2b_prioridad?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cre2b_modulosstatecode;
  statuscode?: Cre2b_modulosstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cre2b_modulos extends Cre2b_modulosBase {
  cre2b_aplicacionname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cre2b_aplicacion?: object;
  _cre2b_aplicacion_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}