// src/generated/models/Cre2b_rolsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cre2b_rolsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cre2b_rolsstatecode = keyof typeof Cre2b_rolsstatecode;
export const Cre2b_rolsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cre2b_rolsstatuscode = keyof typeof Cre2b_rolsstatuscode;

export interface Cre2b_rolsBase {
  cre2b_alias?: string;
  cre2b_nombre?: string;
  cre2b_rolid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cre2b_rolsstatecode;
  statuscode?: Cre2b_rolsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cre2b_rols extends Cre2b_rolsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}