// src/generated/models/Cre2b_modulos_rolesesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cre2b_modulos_rolesesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cre2b_modulos_rolesesstatecode = keyof typeof Cre2b_modulos_rolesesstatecode;
export const Cre2b_modulos_rolesesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cre2b_modulos_rolesesstatuscode = keyof typeof Cre2b_modulos_rolesesstatuscode;

export interface Cre2b_modulos_rolesesBase {
  "cre2b_Modulo@odata.bind"?: string;
  cre2b_modulos_roles1?: string;
  cre2b_modulos_rolesid: string;
  "cre2b_Rol@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cre2b_modulos_rolesesstatecode;
  statuscode?: Cre2b_modulos_rolesesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cre2b_modulos_roleses extends Cre2b_modulos_rolesesBase {
  cre2b_moduloname?: string;
  cre2b_rolname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cre2b_modulo?: object;
  _cre2b_modulo_value?: string;
  cre2b_rol?: object;
  _cre2b_rol_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}