// src/generated/models/Cre2b_usuarios_rolesesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cre2b_usuarios_rolesesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cre2b_usuarios_rolesesstatecode = keyof typeof Cre2b_usuarios_rolesesstatecode;
export const Cre2b_usuarios_rolesesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cre2b_usuarios_rolesesstatuscode = keyof typeof Cre2b_usuarios_rolesesstatuscode;

export interface Cre2b_usuarios_rolesesBase {
  "cre2b_Rol@odata.bind"?: string;
  "cre2b_Usuario@odata.bind"?: string;
  cre2b_usuarios_roles1?: string;
  cre2b_usuarios_rolesid: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cre2b_usuarios_rolesesstatecode;
  statuscode?: Cre2b_usuarios_rolesesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cre2b_usuarios_roleses extends Cre2b_usuarios_rolesesBase {
  cre2b_rolname?: string;
  cre2b_usuarioname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  cre2b_rol?: object;
  _cre2b_rol_value?: string;
  cre2b_usuario?: object;
  _cre2b_usuario_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}